src/requirements.txt
Django>=2.0
beautifulsoup4
lxml
requests
django-taggit